Cache images atomically
The original code could run into a race condition if two processes
wanted to use the same image at the same time. The first process would
start extracting the rpmdb from it, and the second could see partial
data being available in the cache. It could either error out or proceed
as if the image had no RPMs whatsover.